+++ aimer/auto_aim/predictor/enemy_model/motion/spin_motion.cpp
@@ -107,7 +107,7 @@ void SpinMotion::update(const ArmorData& armor, double timestamp) {
 
     // 读取门限参数
     double chi2_threshold = runtime_param::get_param<double>("AutoAim.Predictor.SpinEKF.Gating.chi2_threshold");
-    int max_reject = runtime_param::get_param<int>("AutoAim.Predictor.SpinEKF.Gating.max_reject");
+    int64_t max_reject = runtime_param::get_param<int64_t>("AutoAim.Predictor.SpinEKF.Gating.max_reject");
     double q_scale_increase = runtime_param::get_param<double>("AutoAim.Predictor.SpinEKF.Gating.q_scale_increase");
     double q_scale_decay = runtime_param::get_param<double>("AutoAim.Predictor.SpinEKF.Gating.q_scale_decay");
 
@@ -233,7 +233,7 @@ void SpinMotion::update(const std::vector<ArmorData>& armors, double timestamp) 
 
     // 读取门限参数
     double chi2_threshold = runtime_param::get_param<double>("AutoAim.Predictor.SpinEKF.Gating.chi2_threshold");
-    int max_reject = runtime_param::get_param<int>("AutoAim.Predictor.SpinEKF.Gating.max_reject");
+    int64_t max_reject = runtime_param::get_param<int64_t>("AutoAim.Predictor.SpinEKF.Gating.max_reject");
     double q_scale_increase = runtime_param::get_param<double>("AutoAim.Predictor.SpinEKF.Gating.q_scale_increase");
     double q_scale_decay = runtime_param::get_param<double>("AutoAim.Predictor.SpinEKF.Gating.q_scale_decay");
